breaking up low level storage into logical files Signed-off-by: David Lawrence <david.lawrence@docker.com> (github: endophage)
